kafka/requirements.txt
confluent-kafka
grpcio
google-cloud
bigquery
fastavro
greenlet==2.0.1
numpy==1.24.0
pandas==1.3.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
avro-python3==1.10.2
